podcast_generator/requirements.txt
anthropic>=0.39.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
TTS>=0.22.0
pydub>=0.25.1
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
